+++ NEWS.md
@@ -1,6 +1,7 @@
 # neighbourhoodstudy 0.6.0.9000
 
 * Added query_osm_api() function
+* Removed ottawa_phhs_shp from ISED data in preference of ottawa_phhs, custom set
 
 # neighbourhoodstudy 0.5.0