+++ NEWS.md
@@ -1,7 +1,8 @@
-# neighbourhoodstudy 0.6.0.9000
+# neighbourhoodstudy 0.6.0
 
 * Added query_osm_api() function
 * Removed ottawa_phhs_shp from ISED data in preference of ottawa_phhs, custom set
+* Added sli_das_gen3_mape, single-link indicator from 2021 DAs to Gen3 hoods using `sliopt` package and mean average percentage error.
 
 # neighbourhoodstudy 0.5.0